Split monolithic views.py into per-resource routers
Break the 1254-line views.py into a views/ package: sessions.py,
agents.py, environments.py, health.py, plus _shared.py for cross-cutting
Sprites/runtime-key helpers. urls.py now imports the submodules directly.

Update test patches from agent_on_demand.views.* to
agent_on_demand.views.sessions.* since _get_client and threading.Thread
are now session-only.
