+++ src/agent_on_demand/views/agents.py
@@ -1,0 +1,384 @@
+import json
+import re
+
+from django.http import JsonResponse
+from django.utils import timezone
+from django.views.decorators.csrf import csrf_exempt
+from django.views.decorators.http import require_GET, require_POST
+
+from pydantic import BaseModel, Field, ValidationError, field_validator
+
+from agent_on_demand.auth import require_api_key
+from agent_on_demand.models import Agent, AgentVersion, Environment
+from agent_on_demand.runtimes import RUNTIMES, AgentModel
+
+
+AGENT_VERSIONED_FIELDS = (
+    "name",
+    "description",
+    "system",
+    "model",
+    "runtime",
+    "environment",
+    "skills",
+    "mcp_servers",
+    "metadata",
+)
+
+
+VALID_MCP_SERVER_TYPES = {"url", "stdio"}
+
+
+MAX_SKILLS_PER_AGENT = 20
+MAX_SKILL_DESCRIPTION_LEN = 1024
+MAX_SKILL_CONTENT_BYTES = 64 * 1024
+
+_SKILL_NAME_RE = re.compile(r"^[a-z0-9][a-z0-9-]{0,63}$")
+_SKILL_ALLOWED_KEYS = {"name", "description", "content"}
+_SKILL_HEREDOC_DELIMITER = "SKILL_EOF"
+
+
+def _validate_skills(skills: list) -> list:
+    if len(skills) > MAX_SKILLS_PER_AGENT:
+        raise ValueError(f"Maximum {MAX_SKILLS_PER_AGENT} skills per agent")
+    seen_names: set[str] = set()
+    for i, skill in enumerate(skills):
+        if not isinstance(skill, dict):
+            raise ValueError(f"skills[{i}] must be an object")
+
+        extra = set(skill) - _SKILL_ALLOWED_KEYS
+        if extra:
+            raise ValueError(
+                f"skills[{i}]: unknown keys {sorted(extra)!r}. "
+                f"Allowed: {sorted(_SKILL_ALLOWED_KEYS)}"
+            )
+        for field_name in ("name", "description", "content"):
+            if field_name not in skill:
+                raise ValueError(f"skills[{i}] missing required field: {field_name}")
+            if not isinstance(skill[field_name], str):
+                raise ValueError(f"skills[{i}].{field_name} must be a string")
+
+        name = skill["name"]
+        if not _SKILL_NAME_RE.match(name):
+            raise ValueError(f"skills[{i}].name {name!r} must match [a-z0-9][a-z0-9-]{{0,63}}")
+        if name in seen_names:
+            raise ValueError(f"skills[{i}]: duplicate name {name!r}")
+        seen_names.add(name)
+
+        if len(skill["description"]) > MAX_SKILL_DESCRIPTION_LEN:
+            raise ValueError(f"skills[{i}].description exceeds {MAX_SKILL_DESCRIPTION_LEN} chars")
+
+        content = skill["content"]
+        if len(content.encode("utf-8")) > MAX_SKILL_CONTENT_BYTES:
+            raise ValueError(f"skills[{i}].content exceeds {MAX_SKILL_CONTENT_BYTES} bytes")
+        if _SKILL_HEREDOC_DELIMITER in content:
+            raise ValueError(f"skills[{i}].content must not contain {_SKILL_HEREDOC_DELIMITER!r}")
+    return skills
+
+
+def _validate_mcp_servers(servers: list) -> list:
+    names = set()
+    for i, server in enumerate(servers):
+        if not isinstance(server, dict):
+            raise ValueError(f"mcp_servers[{i}] must be an object")
+        if "name" not in server:
+            raise ValueError(f"mcp_servers[{i}] missing required field: name")
+        stype = server.get("type", "url")
+        if stype not in VALID_MCP_SERVER_TYPES:
+            raise ValueError(
+                f"mcp_servers[{i}]: unknown type {stype!r}. "
+                f"Must be one of: {sorted(VALID_MCP_SERVER_TYPES)}"
+            )
+        if stype == "url" and "url" not in server:
+            raise ValueError(f"mcp_servers[{i}] (url): missing required field: url")
+        if stype == "stdio" and "command" not in server:
+            raise ValueError(f"mcp_servers[{i}] (stdio): missing required field: command")
+        if server["name"] in names:
+            raise ValueError(f"mcp_servers[{i}]: duplicate name {server['name']!r}")
+        names.add(server["name"])
+    if len(servers) > 20:
+        raise ValueError("Maximum 20 MCP servers per agent")
+    return servers
+
+
+class CreateAgentRequest(BaseModel):
+    name: str = Field(max_length=200)
+    model: str = Field(max_length=100)
+    runtime: str = Field(max_length=32)
+    system: str = Field(default="")
+    description: str = Field(default="")
+    environment_id: str | None = Field(default=None)
+    skills: list = Field(default_factory=list)
+    mcp_servers: list = Field(default_factory=list)
+    metadata: dict = Field(default_factory=dict)
+
+    @field_validator("model")
+    @classmethod
+    def validate_model(cls, v: str) -> str:
+        if v not in AgentModel.values():
+            raise ValueError(f"Unknown model: {v}. Must be one of: {sorted(AgentModel.values())}")
+        return v
+
+    @field_validator("mcp_servers")
+    @classmethod
+    def validate_mcp_servers(cls, v: list) -> list:
+        return _validate_mcp_servers(v)
+
+    @field_validator("skills")
+    @classmethod
+    def validate_skills(cls, v: list) -> list:
+        return _validate_skills(v)
+
+
+class UpdateAgentRequest(BaseModel):
+    version: int = Field(description="Current version — optimistic concurrency check")
+    name: str | None = None
+    model: str | None = None
+    runtime: str | None = None
+    system: str | None = None
+    description: str | None = None
+    environment_id: str | None = None
+    skills: list | None = None
+    mcp_servers: list | None = None
+    metadata: dict | None = None
+
+    @field_validator("model")
+    @classmethod
+    def validate_model(cls, v: str | None) -> str | None:
+        if v is not None and v not in AgentModel.values():
+            raise ValueError(f"Unknown model: {v}. Must be one of: {sorted(AgentModel.values())}")
+        return v
+
+    @field_validator("mcp_servers")
+    @classmethod
+    def validate_mcp_servers(cls, v: list | None) -> list | None:
+        if v is not None:
+            _validate_mcp_servers(v)
+        return v
+
+    @field_validator("skills")
+    @classmethod
+    def validate_skills(cls, v: list | None) -> list | None:
+        if v is not None:
+            _validate_skills(v)
+        return v
+
+
+def _serialize_agent(agent: Agent) -> dict:
+    return {
+        "id": str(agent.id),
+        "type": "agent",
+        "name": agent.name,
+        "description": agent.description or None,
+        "system": agent.system or None,
+        "model": agent.model,
+        "runtime": agent.runtime,
+        "environment_id": str(agent.environment_id) if agent.environment_id else None,
+        "skills": agent.skills,
+        "mcp_servers": agent.mcp_servers,
+        "metadata": agent.metadata,
+        "version": agent.version,
+        "created_at": agent.created_at.isoformat(),
+        "updated_at": agent.updated_at.isoformat(),
+        "archived_at": agent.archived_at.isoformat() if agent.archived_at else None,
+    }
+
+
+def _serialize_agent_version(av: AgentVersion) -> dict:
+    return {
+        "id": str(av.agent_id),
+        "type": "agent",
+        "name": av.name,
+        "description": av.description or None,
+        "system": av.system or None,
+        "model": av.model,
+        "runtime": av.runtime,
+        "environment_id": str(av.environment_id) if av.environment_id else None,
+        "skills": av.skills,
+        "mcp_servers": av.mcp_servers,
+        "metadata": av.metadata,
+        "version": av.version,
+        "created_at": av.created_at.isoformat(),
+    }
+
+
+def _snapshot_version(agent: Agent):
+    """Save the current agent state as a version record."""
+    AgentVersion.objects.create(
+        agent=agent,
+        version=agent.version,
+        name=agent.name,
+        description=agent.description,
+        system=agent.system,
+        model=agent.model,
+        runtime=agent.runtime,
+        environment=agent.environment,
+        skills=agent.skills,
+        mcp_servers=agent.mcp_servers,
+        metadata=agent.metadata,
+    )
+
+
+@csrf_exempt
+@require_api_key
+def agents_list_create(request):
+    """POST: create agent. GET: list agents."""
+    if request.method == "POST":
+        try:
+            body = json.loads(request.body)
+        except (json.JSONDecodeError, ValueError):
+            return JsonResponse({"detail": "Invalid JSON"}, status=400)
+
+        try:
+            req = CreateAgentRequest(**body)
+        except ValidationError as e:
+            return JsonResponse({"detail": e.errors(include_context=False)}, status=422)
+
+        if req.runtime not in RUNTIMES:
+            return JsonResponse(
+                {"detail": f"Unknown runtime: {req.runtime}. Must be one of: {list(RUNTIMES)}"},
+                status=400,
+            )
+
+        env_obj = None
+        if req.environment_id:
+            try:
+                env_obj = Environment.objects.get(pk=req.environment_id, user=request.user)
+            except (Environment.DoesNotExist, ValueError):
+                return JsonResponse({"detail": "Environment not found"}, status=404)
+
+        agent = Agent.objects.create(
+            user=request.user,
+            name=req.name,
+            description=req.description,
+            system=req.system,
+            model=req.model,
+            runtime=req.runtime,
+            environment=env_obj,
+            skills=req.skills,
+            mcp_servers=req.mcp_servers,
+            metadata=req.metadata,
+            version=1,
+        )
+        _snapshot_version(agent)
+
+        return JsonResponse(_serialize_agent(agent), status=201)
+
+    elif request.method == "GET":
+        qs = Agent.objects.filter(user=request.user, archived_at__isnull=True).order_by(
+            "-created_at"
+        )
+        return JsonResponse({"data": [_serialize_agent(a) for a in qs]})
+
+    return JsonResponse({"detail": "Method not allowed"}, status=405)
+
+
+@csrf_exempt
+@require_api_key
+def agent_detail(request, agent_id):
+    """GET: retrieve agent. PUT: update agent."""
+    try:
+        agent = Agent.objects.get(pk=agent_id, user=request.user)
+    except (Agent.DoesNotExist, ValueError):
+        return JsonResponse({"detail": "Agent not found"}, status=404)
+
+    if request.method == "GET":
+        return JsonResponse(_serialize_agent(agent))
+
+    if request.method == "PUT":
+        if agent.is_archived:
+            return JsonResponse({"detail": "Cannot update an archived agent"}, status=409)
+
+        try:
+            body = json.loads(request.body)
+        except (json.JSONDecodeError, ValueError):
+            return JsonResponse({"detail": "Invalid JSON"}, status=400)
+
+        try:
+            req = UpdateAgentRequest(**body)
+        except ValidationError as e:
+            return JsonResponse({"detail": e.errors(include_context=False)}, status=422)
+
+        if req.version != agent.version:
+            return JsonResponse(
+                {"detail": f"Version mismatch: expected {agent.version}, got {req.version}"},
+                status=409,
+            )
+
+        if req.runtime is not None and req.runtime not in RUNTIMES:
+            return JsonResponse(
+                {"detail": f"Unknown runtime: {req.runtime}. Must be one of: {list(RUNTIMES)}"},
+                status=400,
+            )
+
+        # Detect changes
+        changed = False
+
+        # Resolve environment_id if provided
+        if req.environment_id is not None:
+            try:
+                env_obj = Environment.objects.get(pk=req.environment_id, user=request.user)
+            except (Environment.DoesNotExist, ValueError):
+                return JsonResponse({"detail": "Environment not found"}, status=404)
+            if env_obj.id != agent.environment_id:
+                agent.environment = env_obj
+                changed = True
+
+        for field in ("name", "model", "runtime", "system", "description", "skills", "mcp_servers"):
+            value = getattr(req, field)
+            if value is not None and value != getattr(agent, field):
+                setattr(agent, field, value)
+                changed = True
+
+        # Metadata merges at key level (matching Anthropic semantics)
+        if req.metadata is not None:
+            merged = dict(agent.metadata)
+            for k, v in req.metadata.items():
+                if v == "":
+                    merged.pop(k, None)
+                else:
+                    merged[k] = v
+            if merged != agent.metadata:
+                agent.metadata = merged
+                changed = True
+
+        if changed:
+            agent.version += 1
+            agent.save()
+            _snapshot_version(agent)
+
+        return JsonResponse(_serialize_agent(agent))
+
+    return JsonResponse({"detail": "Method not allowed"}, status=405)
+
+
+@csrf_exempt
+@require_POST
+@require_api_key
+def agent_archive(request, agent_id):
+    """Archive an agent (read-only, no new sessions)."""
+    try:
+        agent = Agent.objects.get(pk=agent_id, user=request.user)
+    except (Agent.DoesNotExist, ValueError):
+        return JsonResponse({"detail": "Agent not found"}, status=404)
+
+    if agent.is_archived:
+        return JsonResponse({"detail": "Agent is already archived"}, status=409)
+
+    agent.archived_at = timezone.now()
+    agent.save(update_fields=["archived_at", "updated_at"])
+
+    return JsonResponse(_serialize_agent(agent))
+
+
+@require_GET
+@require_api_key
+def agent_versions(request, agent_id):
+    """List all versions of an agent."""
+    try:
+        agent = Agent.objects.get(pk=agent_id, user=request.user)
+    except (Agent.DoesNotExist, ValueError):
+        return JsonResponse({"detail": "Agent not found"}, status=404)
+
+    versions = AgentVersion.objects.filter(agent=agent).order_by("-version")
+    return JsonResponse({"data": [_serialize_agent_version(av) for av in versions]})
