+++ src/agent_on_demand/views/health.py
@@ -1,0 +1,7 @@
+from django.http import JsonResponse
+from django.views.decorators.http import require_GET
+
+
+@require_GET
+def health(request):
+    return JsonResponse({"status": "ok"})
