+++ tests/test_skills.py
@@ -72,8 +72,8 @@ def mock_sprites(mocker):
     mock_sprite.command.return_value.run = mocker.Mock()
     mock_client = mocker.MagicMock()
     mock_client.create_sprite.return_value = mock_sprite
-    mocker.patch("agent_on_demand.views._get_client", return_value=mock_client)
-    mocker.patch("agent_on_demand.views.threading.Thread")
+    mocker.patch("agent_on_demand.views.sessions._get_client", return_value=mock_client)
+    mocker.patch("agent_on_demand.views.sessions.threading.Thread")
     return mock_sprite, mock_fs
 
 
@@ -403,7 +403,7 @@ class TestSessionSkillsIntegration:
         )
         mock_sprite, mock_fs = mock_sprites
         mocker.patch(
-            "agent_on_demand.views._get_client",
+            "agent_on_demand.views.sessions._get_client",
             return_value=mocker.MagicMock(get_sprite=mocker.Mock(return_value=mock_sprite)),
         )
         resp = client.post(
